refactor: preference controller, service 삭제 후 member controller, service 생성 (#14)
